serverless: rename classes back to original names
